Color changing
Two changes in this code.
The first is a separation of the color of the tab from the function of
the tab.
The second is dynamic changing from one color scheme to another.

The visual indication of input with "remembered" (using cookies) data is
still stuck because there is no good way to apply styles to a checkbox.
I have found or tested many many bad ways.